ai_assistant/requirements.txt
# Core Framework
fastapi==0.95.2
uvicorn==0.22.0
python-multipart==0.0.6

# Database
sqlalchemy==1.4.53
alembic==1.11.1

# External APIs
twilio==8.1.0
requests==2.31.0
aiohttp==3.8.5

# Scheduling
apscheduler==3.10.1

# Data Processing
pandas==1.5.3
openpyxl==3.1.2
xlsxwriter==3.1.9

# PDF Generation
fpdf2==2.7.6

# Encryption & Security
cryptography==39.0.2
python-dotenv==1.0.0

# Email
email-validator==2.0.0

# Utilities
python-dateutil==2.8.2
pytz==2023.3
click==8.1.3
jinja2==3.1.2

# Testing
pytest==7.3.1
pytest-asyncio==0.21.0

# Development
black==23.3.0
flake8==5.0.4

# Monitoring & Logging
structlog==23.1.0

# Environment
python-decouple==3.8 